requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
coinpayments-py==0.1.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-coinpayments==0.1.3
django-model-utils==4.1.1
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
platformdirs==2.2.0
psycopg2-binary==2.9.1
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
six==1.15.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.6.0
whitenoise==5.3.0
